requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-debug-toolbar==6.0.0
django-extensions==4.1
django-jazzmin==3.0.1
mysql-connector-python==9.4.0
mysqlclient==2.2.7
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
ruff==0.12.11
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.9.0
